requirements.txt
Django==5.2.3
djangorestframework==3.14.0
django-filter==23.5
flake8==7.0.0
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
python-decouple==3.8 